backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# Uncomment when activating LLM features:
# openai==1.30.1
